requirements.txt
django
djangorestframework
django-cors-headers
requests
djongo
djangorestframework-simplejwt